Dirty-merge homestead origin changes
https://github.com/laravel/homestead/tree/69e1e46b5b815953d8c1860b7dc6c647fe874afe

+++ scripts/homestead.rb
@@ -40,7 +40,7 @@ class Homestead
     # Configure A Few VMware Settings
     ["vmware_fusion", "vmware_workstation"].each do |vmware|
       config.vm.provider vmware do |v|
-        v.vmx["displayName"] = "homestead"
+        v.vmx["displayName"] = settings["name"] ||= "homestead-7"
         v.vmx["memsize"] = settings["memory"] ||= 2048
         v.vmx["numvcpus"] = settings["cpus"] ||= 1
         v.vmx["guestOS"] = "ubuntu-64"
@@ -74,9 +74,11 @@ class Homestead
     }
 
     # Use Default Port Forwarding Unless Overridden
-    default_ports.each do |guest, host|
-      unless settings["ports"].any? { |mapping| mapping["guest"] == guest }
-        config.vm.network "forwarded_port", guest: guest, host: host, auto_correct: true
+    unless settings.has_key?("default_ports") && settings["default_ports"] == false
+      default_ports.each do |guest, host|
+        unless settings["ports"].any? { |mapping| mapping["guest"] == guest }
+          config.vm.network "forwarded_port", guest: guest, host: host, auto_correct: true
+        end
       end
     end
 
@@ -108,6 +110,16 @@ class Homestead
       end
     end
 
+    # Copy User Files Over to VM
+    if settings.include? 'copy'
+      settings["copy"].each do |file|
+        config.vm.provision "file" do |f|
+          f.source = File.expand_path(file["from"])
+          f.destination = file["to"].chomp('/') + "/" + file["from"].split('/').last
+        end
+      end
+    end
+
     # Register All Of The Configured Shared Folders
     if settings.include? 'folders'
       settings["folders"].each do |folder|
@@ -115,6 +127,8 @@ class Homestead
 
         if (folder["type"] == "nfs")
             mount_opts = folder["mount_options"] ? folder["mount_options"] : ['actimeo=1']
+        elsif (folder["type"] == "smb")
+            mount_opts = folder["mount_options"] ? folder["mount_options"] : ['vers=3.02', 'mfsymlinks']
         end
 
         # For b/w compatibility keep separate 'mount_opts', but merge with options
@@ -124,6 +138,11 @@ class Homestead
         options.keys.each{|k| options[k.to_sym] = options.delete(k) }
 
         config.vm.synced_folder folder["map"], folder["to"], type: folder["type"] ||= nil, **options
+
+        # Bindfs support to fix shared folder (NFS) permission issue on Mac
+        if Vagrant.has_plugin?("vagrant-bindfs")
+          config.bindfs.bind_folder folder["to"], folder["to"]
+        end
       end
     end
 
@@ -133,46 +152,66 @@ class Homestead
     end
 
 
-    settings["sites"].each do |site|
-      type = site["type"] ||= "laravel"
+    if settings.include? 'sites'
+      settings["sites"].each do |site|
+        type = site["type"] ||= "laravel"
 
-      if (site.has_key?("hhvm") && site["hhvm"])
-        type = "hhvm"
-      end
+        if (site.has_key?("hhvm") && site["hhvm"])
+          type = "hhvm"
+        end
 
-      if (type == "symfony")
-        type = "symfony2"
-      end
+        if (type == "symfony")
+          type = "symfony2"
+        end
 
-      config.vm.provision "shell" do |s|
-        s.path = scriptDir + "/serve-#{type}.sh"
-        s.args = [site["map"], site["to"], site["port"] ||= "80", site["ssl"] ||= "443"]
-      end
-
-      # Configure The Cron Schedule
-      if (site.has_key?("schedule"))
         config.vm.provision "shell" do |s|
-          if (site["schedule"])
-            s.path = scriptDir + "/cron-schedule.sh"
-            s.args = [site["map"].tr('^A-Za-z0-9', ''), site["to"]]
-          else
-            s.inline = "rm -f /etc/cron.d/$1"
-            s.args = [site["map"].tr('^A-Za-z0-9', '')]
+          s.name = "Creating Site: " + site["map"]
+          s.path = scriptDir + "/serve-#{type}.sh"
+          s.args = [site["map"], site["to"], site["port"] ||= "80", site["ssl"] ||= "443"]
+        end
+
+        # Configure The Cron Schedule
+        if (site.has_key?("schedule"))
+          config.vm.provision "shell" do |s|
+            s.name = "Creating Schedule"
+
+            if (site["schedule"])
+              s.path = scriptDir + "/cron-schedule.sh"
+              s.args = [site["map"].tr('^A-Za-z0-9', ''), site["to"]]
+            else
+              s.inline = "rm -f /etc/cron.d/$1"
+              s.args = [site["map"].tr('^A-Za-z0-9', '')]
+            end
           end
         end
-      end
 
+      end
     end
+
+    config.vm.provision "shell" do |s|
+      s.name = "Restarting Nginx"
+      s.inline = "sudo service nginx restart; sudo service php7.0-fpm restart"
+    end
+
+    # Install MariaDB If Necessary
+    if settings.has_key?("mariadb") && settings["mariadb"]
+      config.vm.provision "shell" do |s|
+        s.path = scriptDir + "/install-maria.sh"
+      end
+    end
+
 
     # Configure All Of The Configured Databases
     if settings.has_key?("databases")
         settings["databases"].each do |db|
           config.vm.provision "shell" do |s|
+            s.name = "Creating MySQL Database"
             s.path = scriptDir + "/create-mysql.sh"
             s.args = [db]
           end
 
           config.vm.provision "shell" do |s|
+            s.name = "Creating Postgres Database"
             s.path = scriptDir + "/create-postgres.sh"
             s.args = [db]
           end
@@ -181,6 +220,7 @@ class Homestead
 
     # Configure All Of The Server Environment Variables
     config.vm.provision "shell" do |s|
+        s.name = "Clear Variables"
         s.path = scriptDir + "/clear-variables.sh"
     end
